HID: Fix logitech-dj: Allow incoming packets during probe().
Historically, logitech-dj communicated with the device during probe()
to query the list of devices attached. Later, a change was introduced
to hid-core that prevented incoming packets for a device during
probe(), as many drivers are unable to handle such input. That change
broke the device enumeration in logitech-dj, so commit
596264082f10dd4a56 was introduced to workaround that by waiting for
normal input before enumerating devices.

Now that drivers can opt-in to receive input during probe, this patch
changes logitech-dj to do that, so that it can successfully complete
enumeration of devices during probe().

After this, another patch can go in to revert the previous workaround
(596264082f10dd4a56).

Change-Id: I2dca10a5c64f6c3e90ffb00547b47eaf94546eb6